feat: Add Restaurant Categories and Reservations
The code changes include adding the `Category` and `Reservation` models to the `RestaurantReservationAPI` project. The `RestaurantDbContext` class was modified to include `DbSet` properties for `Reservations` and `Categories`. The `Menu` model was also updated to include `ImageUrl` and `CategoryId` properties. Additionally, several new classes and interfaces were added to support GraphQL queries and mutations for `Categories` and `Reservations`.

Recent user commits:
- feat: Add SqlServer db store and migrations for initial setup
- chore: Add Nuget packages Microsoft.EntityFrameworkCore and Microsoft.EntityFrameworkCore.SqlServer
- Add SQL Server Docker container configuration
- feat: Add mutation operations for Menu using GraphQL
- feat: implement MenuQuery using GraphQL
- chore: Add GraphQL dependencies and project structure
- initialize api project
- Initial commit